+++ addon/components/object-list-view.js
@@ -367,9 +367,9 @@ export default FlexberryBaseComponent.extend(FlexberryLookupCompatibleComponentM
     'modelProjection',
     function() {
     if (this.get('modelProjection')) {
-      return false;
-    } else {
       return this.get('showAsteriskInRow') || this.get('showCheckBoxInRow') || this.get('showDeleteButtonInRow');
+    } else {
+      return false;
     }
   }),
 
@@ -909,6 +909,11 @@ export default FlexberryBaseComponent.extend(FlexberryLookupCompatibleComponentM
     getSettingPromise.then(function(data) {
       _this._setColumnWidths(data);
     });
+
+    // TODO: resolv this problem.
+    this.$('.flexberry-dropdown:last').dropdown({
+      direction: 'upward'
+    });
   },
 
   /**
@@ -917,8 +922,11 @@ export default FlexberryBaseComponent.extend(FlexberryLookupCompatibleComponentM
    *
    * @method didRender
    */
-  didRender: function() {
+  didRender() {
     this._super(...arguments);
+
+    let currentTable = this.$('table.object-list-view');
+
     if (this.get('allowColumnResize')) {
       if (this.get('useSingleColumn')) {
         Ember.Logger.error(
@@ -926,8 +934,6 @@ export default FlexberryBaseComponent.extend(FlexberryLookupCompatibleComponentM
           'can\'t be enabled at the same time.');
         return;
       }
-
-      let currentTable = this.$('table.object-list-view');
 
       // The first column has semantic class "collapsing"
       // so the column has 1px width and plugin has problems.
@@ -940,6 +946,8 @@ export default FlexberryBaseComponent.extend(FlexberryLookupCompatibleComponentM
       currentTable.addClass('fixed');
 
       this._reinitResizablePlugin();
+    } else {
+      currentTable.colResizable({ disable: true });
     }
   },
 
@@ -950,18 +958,18 @@ export default FlexberryBaseComponent.extend(FlexberryLookupCompatibleComponentM
    * @method _reinitResizablePlugin
    * @private
    */
-  _reinitResizablePlugin: function() {
+  _reinitResizablePlugin() {
     let currentTable = this.$('table.object-list-view');
 
     // Disable plugin and then init it again.
     currentTable.colResizable({ disable: true });
 
-    let _this = this;
     currentTable.colResizable({
-      minWidth: 70,
-      onResize: function(e) {
+      minWidth: 90,
+      resizeMode:'flex',
+      onResize: (e)=> {
         // Save column width as user setting on resize.
-        _this._afterColumnResize(e);
+        this._afterColumnResize(e);
       }
     });
   },