+++ CHANGELOG.md
@@ -1,11 +1,16 @@
 # Ember Flexberry Changelog
 ### New Features & improvements
 
-## 2016-06-21
+## 2016-06-23
 ### Added
 FlexberryObjectlistviewComponent
 * Add support `modelProjection` by name
 * Add support promise for content
+
+## 2016-06-22
+### Added
+FlexberryCheckboxComponent:
+* Add class property for wrapper component.
 
 ## 2016-06-14
 ### Fixed